DJ-142: zero type errors on the signal path, and two of them were real defects
99 errors to 32, and every one of the 32 is off the nightly path: reporting,
knowledge graph, training data, EDGAR's network fetch, and two in the untracked
paper modules that are deliberately not touched. hifi/live, hifi/mcp,
hifi/portfolio, hifi/execution, the six agents, hifi/engines,
simulation/{agent_executor,snapshot,model_manager} and collective/voting are
clean.

Two findings that were defects rather than annotations, both on the path a
signal takes from the model to an order:

  BaseMessage.content is `str | list[str | dict]` -- providers may answer with
  content blocks instead of a string. All seven call sites passed it straight
  into extract_json, which begins text.strip(). A block response would have
  raised AttributeError inside the agent's own try/except and been recorded as a
  parse failure naming neither the cause nor the text: the identical shape to
  the list-from-json.loads case fixed at DJ-140, one layer earlier. Local LM
  Studio models return strings, which is a property of the serving stack and not
  of the interface. hifi.agents.json_parsing.message_text now flattens it.

  MCP tool results, LM Studio responses and book_state.json were all read with
  json.loads and handed to callers that immediately do .get(). Each now refuses
  a non-object at the boundary and says what it got.

The rest, in order of how much they were hiding:

  _test_llm was typed `object | None` in all six agents and in agent_executor.
  That joined with the make_llm() branch to `object`, producing three errors per
  agent AND masking the content-union finding underneath. Annotating ChatOpenAI
  would have been untrue -- the injected doubles are fakes, and requiring them to
  be ChatOpenAI makes the seam useless. hifi.agents.lm_client.ChatModel is a
  Protocol for what the nodes actually call: invoke and model_name. ChatOpenAI
  satisfies it structurally; so does any double.

  confidence_weighted_vote took list[AgentSignal | None]. list is invariant, so a
  caller that had already filtered out the Nones could not pass its
  list[AgentSignal]. It reads and never mutates, so the type is Sequence.

  simulation/snapshot: four float() calls on pandas cells, and period_end taking
  a str where FundamentalsSnapshot declares date. Pydantic coerced it -- on the
  field that decides which fundamentals an agent may see. Now
  date.fromisoformat, so a malformed date fails naming itself. _cell_float keeps
  an unreported quarter as None rather than 0.0: a company that reported nothing
  must not look like one that reported nothing of value.

  alpaca-py's Model | dict[str, Any] narrowing moved out of alpaca_executor into
  hifi/execution/alpaca_types, and market_data and news now reuse it instead of
  growing their own.

  make_llm passes SecretStr and keeps max_tokens with a stated ignore. Both were
  verified against the installed langchain-openai (max_tokens=1024 round-trips,
  api_key already coerces). This is the factory every agent's model comes from,
  not somewhere to change kwargs to satisfy a checker.

  .values -> .to_numpy() in engines/{risk,technical,market_summary}: same values,
  the ndarray contract numpy is actually being handed.

yfinance, fredapi, scipy, quantstats and lancedb ship no stubs; declared as
ignore_missing_imports with every crossing cast explicitly at the call site, so
the contract is in code rather than assumed. The notices were burying real
errors.

45 tests for message_text and the ChatModel protocol, including one that pins
the old failure: extract_json on a list still raises.

3,237 passed, 6 skipped, lint clean.

+++ src/hifi/data/regime.py
@@ -65,7 +65,7 @@ def _spy_52w_return(ohlcv: pd.DataFrame, as_of: pd.Timestamp) -> float | None:
     price_year_ago = data.iloc[-_TRADING_YEAR]
     if price_year_ago == 0:
         return None
-    return (price_now - price_year_ago) / price_year_ago
+    return float((price_now - price_year_ago) / price_year_ago)
 
 
 def _close_col(df: pd.DataFrame) -> str:
